resolver auth added for mutations

+++ src/graphql/resolvers/airportResolver.js
@@ -47,7 +47,10 @@ export const AirportMutation = {
       country: { type: GraphQLString },
       locationID: { type: GraphQLString }
     },
-    resolve(parent, args) {
+    resolve(parent, args, {userRoles}) {
+      if (!userRoles.includes('Admin') || userRoles.length === 0) {
+        throw new Error('You are not authorized to do this action');
+      }
       return new Promise((resolve, reject) => {
         db.query('INSERT INTO airport (airportID, airport_name, city, state, country, locationID) VALUES (?, ?, ?, ?, ?, ?)',
           [args.airportID, args.airport_name, args.city, args.state, args.country, args.locationID], 
@@ -74,7 +77,10 @@ export const AirportMutation = {
       country: { type: GraphQLString },
       locationID: { type: GraphQLString }
     },
-    resolve(parent, args) {
+    resolve(parent, args, {userRoles}) {
+      if (!userRoles.includes('Admin') || userRoles.length === 0) {
+        throw new Error('You are not authorized to do this action');
+      }
       const { airportID, ...updateValues } = args;
       const fields = Object.keys(updateValues).filter(key => updateValues[key] != null);
       const values = fields.map(key => updateValues[key]);
@@ -104,7 +110,10 @@ export const AirportMutation = {
     args: {
       airportID: { type: new GraphQLNonNull(GraphQLString) }
     },
-    resolve(parent, args) {
+    resolve(parent, args, {userRoles}) {
+      if (!userRoles.includes('Admin') || userRoles.length === 0) {
+        throw new Error('You are not authorized to do this action');
+      }
       return new Promise((resolve, reject) => {
         db.query('DELETE FROM airport WHERE airportID = ?', [args.airportID], (error) => {
           if (error) {
